refactor: cleanup and make separate the modals
